+++ NEWS.md
@@ -1,6 +1,7 @@
 # taxonbridge 1.0.3
 
-* Fixed bug in get_inconsistencies() when uninomials are set to true
+* Added set operation parameter to `get_inconsistencies()`
+* Fixed bug in `get_inconsistencies()` when uninomials are set to true
 * Added `download_ncbi()` method
 
 # taxonbridge 1.0.2